タスクスイッチができるようになった

+++ src/memory.c
@@ -38,7 +38,7 @@ unsigned int memtest_sub(unsigned int start, unsigned int end)
 {
     unsigned int i, *p, old, pat0 = 0xaa55aa55, pat1 = 0x55aa55aa;
     for (i = start;i <= end;i += 0x1000) {
-        p = (unsigned int *) (i + 0xffc);
+        p = (unsigned int *)(unsigned int)(i + 0xffc);
         old = *p;
         *p = pat0;
         *p ^= 0xffffffff;
